DataStoreService is now able to delete WebDav collections.
The root cause of the problem was in the 'sardine' library not following
HTTP 301 redirects when doing DELETE on resource on WebDav.

In case of GD WebDav infrastructure, when you want to delete collection,
you need to execute DELETE on resource with trailing '/', otherwise you
get HTTP 301 with correct location. Because of that, UriPrefixer was
changed to stop removing trailing slash from given URI.