fix webdav upload 500 response by lowering the 100 (continue) wait timeout
The default timeout in apache HttpClient is 3s. Since the gooddata webdav never sends 100 (continue) response as specification suggests, it causes unfortunate 3s delay in each PUT on webdav. Additionaly it caused 500 when the PUT was unauthenticated. When this timeout is lowered (approx less than 2s) it can work smoothly. So this brings configurable waitForContinue timeout with default of 1s.