Add support for extracting object id for metadata objects.
There are other methods in the SDK that work with object ID
(e.g. MetadataService.getObjById()) and sometimes it may be
useful to remember just an ID of object (rather than the
whole URI) to be able to retrieve it later. However, there
is no straightforward way to get ID of an object (it must
be parsed from the URI). Adding method for obtaining object
ID makes the SDK more complete and is consistent with
other parts of the SDK (ProjectService).