Simplified VisualizationConverter methods for Execution conversion
We wanted to convert objects of VisualizationObject and VisualizationClass that we obtained beforehand, but we were forced to write some boilerplate function that returned the VisualizationClass instead of providing the VisualizationClass directly. The modified methods simply allow to provide the VisualizationClass directly, while still keeping the backward compatibility with previous approach.
In addition, new check was implemented that checks whether the VisualizationClass URI matches with the URI that VisualizationObject contains. If you try to create convert incompatible objects, exception will be thrown.